+++ Cargo.toml
@@ -25,6 +25,7 @@ hyper-util = {version = "0.1.10", features = ["service"]}
 indexmap = {version = "2.8.0", features = ["serde"]}
 itertools = "0.14.0"
 nix = {version = "0.29.0", features = ["process", "signal", "socket", "net"]}
+once_cell = "1.21.3"
 opentelemetry = "0.29.1"
 opentelemetry-otlp = {version = "0.29.0", default-features = false, features = [
   "logs",
@@ -66,10 +67,12 @@ tokio = {version = "1.43.0", features = ["rt-multi-thread", "sync"]}
 tokio-graceful = "0.2.2"
 tokio-rustls = {version = "0.26.1", default-features = false, features = ["logging", "tls12", "ring"]}
 tokio-util = "0.7.13"
+tonic = "0.12.3"
 tower = {version = "0.5", features = ["make", "util"]}
 tower-http = {version = "0.6.2", features = ["trace", "set-header"]}
 tracing = "0.1"
 tracing-log = "0.2.0"
+tracing-opentelemetry = "0.30.0"
 tracing-subscriber = {version = "0.3", features = ["env-filter"]}
 which = "7.0.2"
