+++ src/lib.rs
@@ -1,113 +1,34 @@
-use scopeguard::defer;
 use shadow_rs::shadow;
-use tokio_util::sync::CancellationToken;
-use tunnel::TngRuntime;
-
-use anyhow::{Context as _, Result};
-use config::TngConfig;
 
 pub mod config;
 mod control_interface;
 mod executor;
 mod observability;
-pub mod tunnel;
+pub mod runtime;
+mod service;
+mod state;
+mod tunnel;
 
 shadow!(build);
-
-pub struct TngBuilder {
-    config: TngConfig,
-}
-
-impl TngBuilder {
-    pub fn from_config(mut config: TngConfig) -> Self {
-        if config.admin_bind.is_some() {
-            tracing::warn!("The field `admin_bind` in configuration is ignored, since envoy admin interface is deprecated");
-            config.admin_bind = None;
-        }
-        Self { config }
-    }
-
-    pub async fn serve_forever(self) -> Result<()> {
-        self.serve_with_cancel(CancellationToken::new(), tokio::sync::oneshot::channel().0)
-            .await
-    }
-
-    pub async fn serve_with_cancel(
-        self,
-        cancel_by_caller: CancellationToken, // This is a canell token which can be called from the caller to cancel the task. Note that this funnction will not call the cancel() function on this.
-        ready: tokio::sync::oneshot::Sender<()>,
-    ) -> Result<()> {
-        // Start native part
-        tracing::info!("Starting all service now");
-
-        let runtime = TngRuntime::new_from_config(self.config)
-            .await
-            .context("Failed to launch envoy executor")?;
-
-        let cancel_before_func_return = CancellationToken::new();
-        let for_cancel_safity = cancel_before_func_return.clone();
-        defer! {
-            // Cancel-Safity: exit tng in case of the future of this function is dropped
-            for_cancel_safity.cancel();
-        }
-
-        // Prepare for graceful shutdown
-        let shutdown = {
-            let cancel_before_func_return = cancel_before_func_return.clone();
-            tokio_graceful::Shutdown::builder()
-                .with_signal(async move {
-                    tokio::select! {
-                        _ = cancel_by_caller.cancelled() => {}
-                        _ = cancel_before_func_return.cancelled() => {}
-                        _ = tokio_graceful::default_signal() => {}
-                    }
-                })
-                .with_overwrite_fn(tokio::signal::ctrl_c)
-                .build()
-        };
-
-        // Watch the ready signal from the tng runtime state object.
-        {
-            let mut receiver = runtime.state().ready.0.subscribe();
-            shutdown.spawn_task_fn(move |shutdown_guard| {
-                async move {
-                    loop {
-                        tokio::select! {
-                            _ = receiver.changed() => {
-                                if *receiver.borrow_and_update() {
-                                    let _ = ready.send(());// Ignore any error occuring during send
-                                    break;
-                                }
-                            }
-                            _ = shutdown_guard.cancelled() => {}
-                        }
-                    }
-                }
-            });
-        }
-
-        // Wait for the runtime to finish serving.
-        runtime.serve(shutdown.guard()).await?;
-        // Trigger the shutdown guard to gracefully shutdown all the tokio tasks.
-        cancel_before_func_return.cancel();
-        // Wait for the shutdown guard to complete.
-        shutdown.shutdown().await;
-
-        tracing::debug!("All service shutdown complete");
-        Ok(())
-    }
-}
 
 #[cfg(test)]
 mod tests {
 
+    use anyhow::Result;
+    use once_cell::sync::OnceCell;
+    use scopeguard::defer;
     use serde_json::json;
     use tokio::select;
-    use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
+    use tokio_util::sync::CancellationToken;
+    use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, Layer};
 
-    use super::*;
+    use crate::{
+        config::TngConfig,
+        runtime::{TngRuntime, TracingReloadHandle},
+    };
 
-    #[cfg(test)]
+    pub static RELOAD_HANDLE: OnceCell<TracingReloadHandle> = OnceCell::new();
+
     #[ctor::ctor]
     fn init() {
         // Initialize rustls crypto provider
@@ -116,13 +37,21 @@ mod tests {
             .expect("Failed to install rustls crypto provider");
 
         // Initialize log tracing
-        tracing_subscriber::registry()
-            .with(
+        let pending_tracing_layers = vec![tracing_subscriber::fmt::layer()
+            .with_filter(
                 tracing_subscriber::EnvFilter::try_from_default_env()
-                    .unwrap_or_else(|_| "none,tng=info".into()),
+                    .unwrap_or_else(|_| "info,tng=debug".into()),
             )
-            .with(tracing_subscriber::fmt::layer())
+            .boxed()];
+        let (pending_tracing_layers, reload_handle) =
+            tracing_subscriber::reload::Layer::new(pending_tracing_layers);
+        tracing_subscriber::registry()
+            .with(pending_tracing_layers)
             .init();
+        // Set the reload handle to the global static variable so that we can use it in tests
+        if RELOAD_HANDLE.set(reload_handle).is_err() {
+            panic!("Failed to set reload handle to global static variable")
+        }
     }
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 10)]
@@ -157,7 +86,8 @@ mod tests {
 
         let cancel_token_clone = cancel_token.clone();
         let join_handle = tokio::task::spawn(async move {
-            TngBuilder::from_config(config)
+            TngRuntime::from_config(config)
+                .await?
                 .serve_with_cancel(cancel_token_clone, ready_sender)
                 .await
         });
@@ -214,7 +144,8 @@ mod tests {
 
         let cancel_token_clone = cancel_token.clone();
         let join_handle = tokio::task::spawn(async move {
-            TngBuilder::from_config(config)
+            TngRuntime::from_config(config)
+                .await?
                 .serve_with_cancel(cancel_token_clone, ready_sender)
                 .await
         });
