+++ src/observability/exporter/stdout.rs
@@ -20,7 +20,7 @@ impl SimpleMetricExporter for StdoutExporter {
 #[cfg(test)]
 mod tests {
 
-    use crate::{config::TngConfig, TngBuilder};
+    use crate::{config::TngConfig, runtime::TngRuntime};
     use scopeguard::defer;
     use serde_json::json;
     use tokio::select;
@@ -60,7 +60,8 @@ mod tests {
 
         let cancel_token_clone = cancel_token.clone();
         let join_handle = tokio::task::spawn(async move {
-            TngBuilder::from_config(config)
+            TngRuntime::from_config(config)
+                .await?
                 .serve_with_cancel(cancel_token_clone, ready_sender)
                 .await
         });
