+++ src/service.rs
@@ -1,0 +1,9 @@
+use anyhow::Result;
+use async_trait::async_trait;
+use tokio::sync::mpsc::Sender;
+use tokio_graceful::ShutdownGuard;
+
+#[async_trait]
+pub trait RegistedService {
+    async fn serve(&self, shutdown_guard: ShutdownGuard, ready: Sender<()>) -> Result<()>;
+}
