+++ src/tunnel/egress/mapping/mod.rs
@@ -13,13 +13,13 @@ use tracing::Instrument;
 use crate::{
     config::egress::{CommonArgs, EgressMappingArgs},
     observability::metric::stream::StreamWithCounter,
+    service::RegistedService,
     tunnel::{
         access_log::AccessLog,
         egress::core::stream_manager::{trusted::TrustedStreamManager, StreamManager},
         ingress::core::TngEndpoint,
         service_metrics::ServiceMetrics,
         utils::{self, socket::SetListenerCommonSockOpts},
-        RegistedService,
     },
 };
